CreateShapeRequest Object

CreateShapeRequest Object

Usage

CreateShapeRequest(objectId = NULL, elementProperties = NULL, shapeType = NULL)

Arguments

objectId

(string) A user-supplied object ID. If you specify an ID, it must be unique among all pages and page elements in the presentation. The ID must start with an alphanumeric character or an underscore (matches regex [a-zA-Z0-9_]); remaining characters may include those as well as a hyphen or colon (matches regex [a-zA-Z0-9_-:]). The length of the ID must not be less than 5 or greater than 50. If empty, a unique identifier will be generated.

elementProperties

(PageElementProperties) The element properties for the shape.

shapeType

(string) The shape type.

TYPE_UNSPECIFIED: The shape type that is not predefined.

TEXT_BOX: Text box shape.

RECTANGLE: Rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'rect'.

ROUND_RECTANGLE: Round corner rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'roundRect'

ELLIPSE: Ellipse shape. Corresponds to ECMA-376 ST_ShapeType 'ellipse'

ARC: Curved arc shape. Corresponds to ECMA-376 ST_ShapeType 'arc'

BENT_ARROW: Bent arrow shape. Corresponds to ECMA-376 ST_ShapeType 'bentArrow'

BENT_UP_ARROW: Bent up arrow shape. Corresponds to ECMA-376 ST_ShapeType 'bentUpArrow'

BEVEL: Bevel shape. Corresponds to ECMA-376 ST_ShapeType 'bevel'

BLOCK_ARC: Block arc shape. Corresponds to ECMA-376 ST_ShapeType 'blockArc'

BRACE_PAIR: Brace pair shape. Corresponds to ECMA-376 ST_ShapeType 'bracePair'

BRACKET_PAIR: Bracket pair shape. Corresponds to ECMA-376 ST_ShapeType 'bracketPair'

CAN: Can shape. Corresponds to ECMA-376 ST_ShapeType 'can'

CHEVRON: Chevron shape. Corresponds to ECMA-376 ST_ShapeType 'chevron'

CHORD: Chord shape. Corresponds to ECMA-376 ST_ShapeType 'chord'

CLOUD: Cloud shape. Corresponds to ECMA-376 ST_ShapeType 'cloud'

CORNER: Corner shape. Corresponds to ECMA-376 ST_ShapeType 'corner'

CUBE: Cube shape. Corresponds to ECMA-376 ST_ShapeType 'cube'

CURVED_DOWN_ARROW: Curved down arrow shape. Corresponds to ECMA-376 ST_ShapeType 'curvedDownArrow'

CURVED_LEFT_ARROW: Curved left arrow shape. Corresponds to ECMA-376 ST_ShapeType 'curvedLeftArrow'

CURVED_RIGHT_ARROW: Curved right arrow shape. Corresponds to ECMA-376 ST_ShapeType 'curvedRightArrow'

CURVED_UP_ARROW: Curved up arrow shape. Corresponds to ECMA-376 ST_ShapeType 'curvedUpArrow'

DECAGON: Decagon shape. Corresponds to ECMA-376 ST_ShapeType 'decagon'

DIAGONAL_STRIPE: Diagonal stripe shape. Corresponds to ECMA-376 ST_ShapeType 'diagStripe'

DIAMOND: Diamond shape. Corresponds to ECMA-376 ST_ShapeType 'diamond'

DODECAGON: Dodecagon shape. Corresponds to ECMA-376 ST_ShapeType 'dodecagon'

DONUT: Donut shape. Corresponds to ECMA-376 ST_ShapeType 'donut'

DOUBLE_WAVE: Double wave shape. Corresponds to ECMA-376 ST_ShapeType 'doubleWave'

DOWN_ARROW: Down arrow shape. Corresponds to ECMA-376 ST_ShapeType 'downArrow'

DOWN_ARROW_CALLOUT: Callout down arrow shape. Corresponds to ECMA-376 ST_ShapeType 'downArrowCallout'

FOLDED_CORNER: Folded corner shape. Corresponds to ECMA-376 ST_ShapeType 'foldedCorner'

FRAME: Frame shape. Corresponds to ECMA-376 ST_ShapeType 'frame'

HALF_FRAME: Half frame shape. Corresponds to ECMA-376 ST_ShapeType 'halfFrame'

HEART: Heart shape. Corresponds to ECMA-376 ST_ShapeType 'heart'

HEPTAGON: Heptagon shape. Corresponds to ECMA-376 ST_ShapeType 'heptagon'

HEXAGON: Hexagon shape. Corresponds to ECMA-376 ST_ShapeType 'hexagon'

HOME_PLATE: Home plate shape. Corresponds to ECMA-376 ST_ShapeType 'homePlate'

HORIZONTAL_SCROLL: Horizontal scroll shape. Corresponds to ECMA-376 ST_ShapeType 'horizontalScroll'

IRREGULAR_SEAL_1: Irregular seal 1 shape. Corresponds to ECMA-376 ST_ShapeType 'irregularSeal1'

IRREGULAR_SEAL_2: Irregular seal 2 shape. Corresponds to ECMA-376 ST_ShapeType 'irregularSeal2'

LEFT_ARROW: Left arrow shape. Corresponds to ECMA-376 ST_ShapeType 'leftArrow'

LEFT_ARROW_CALLOUT: Callout left arrow shape. Corresponds to ECMA-376 ST_ShapeType 'leftArrowCallout'

LEFT_BRACE: Left brace shape. Corresponds to ECMA-376 ST_ShapeType 'leftBrace'

LEFT_BRACKET: Left bracket shape. Corresponds to ECMA-376 ST_ShapeType 'leftBracket'

LEFT_RIGHT_ARROW: Left right arrow shape. Corresponds to ECMA-376 ST_ShapeType 'leftRightArrow'

LEFT_RIGHT_ARROW_CALLOUT: Callout left right arrow shape. Corresponds to ECMA-376 ST_ShapeType 'leftRightArrowCallout'

LEFT_RIGHT_UP_ARROW: Left right up arrow shape. Corresponds to ECMA-376 ST_ShapeType 'leftRightUpArrow'

LEFT_UP_ARROW: Left up arrow shape. Corresponds to ECMA-376 ST_ShapeType 'leftUpArrow'

LIGHTNING_BOLT: Lightning bolt shape. Corresponds to ECMA-376 ST_ShapeType 'lightningBolt'

MATH_DIVIDE: Divide math shape. Corresponds to ECMA-376 ST_ShapeType 'mathDivide'

MATH_EQUAL: Equal math shape. Corresponds to ECMA-376 ST_ShapeType 'mathEqual'

MATH_MINUS: Minus math shape. Corresponds to ECMA-376 ST_ShapeType 'mathMinus'

MATH_MULTIPLY: Multiply math shape. Corresponds to ECMA-376 ST_ShapeType 'mathMultiply'

MATH_NOT_EQUAL: Not equal math shape. Corresponds to ECMA-376 ST_ShapeType 'mathNotEqual'

MATH_PLUS: Plus math shape. Corresponds to ECMA-376 ST_ShapeType 'mathPlus'

MOON: Moon shape. Corresponds to ECMA-376 ST_ShapeType 'moon'

NO_SMOKING: No smoking shape. Corresponds to ECMA-376 ST_ShapeType 'noSmoking'

NOTCHED_RIGHT_ARROW: Notched right arrow shape. Corresponds to ECMA-376 ST_ShapeType 'notchedRightArrow'

OCTAGON: Octagon shape. Corresponds to ECMA-376 ST_ShapeType 'octagon'

PARALLELOGRAM: Parallelogram shape. Corresponds to ECMA-376 ST_ShapeType 'parallelogram'

PENTAGON: Pentagon shape. Corresponds to ECMA-376 ST_ShapeType 'pentagon'

PIE: Pie shape. Corresponds to ECMA-376 ST_ShapeType 'pie'

PLAQUE: Plaque shape. Corresponds to ECMA-376 ST_ShapeType 'plaque'

PLUS: Plus shape. Corresponds to ECMA-376 ST_ShapeType 'plus'

QUAD_ARROW: Quad-arrow shape. Corresponds to ECMA-376 ST_ShapeType 'quadArrow'

QUAD_ARROW_CALLOUT: Callout quad-arrow shape. Corresponds to ECMA-376 ST_ShapeType 'quadArrowCallout'

RIBBON: Ribbon shape. Corresponds to ECMA-376 ST_ShapeType 'ribbon'

RIBBON_2: Ribbon 2 shape. Corresponds to ECMA-376 ST_ShapeType 'ribbon2'

RIGHT_ARROW: Right arrow shape. Corresponds to ECMA-376 ST_ShapeType 'rightArrow'

RIGHT_ARROW_CALLOUT: Callout right arrow shape. Corresponds to ECMA-376 ST_ShapeType 'rightArrowCallout'

RIGHT_BRACE: Right brace shape. Corresponds to ECMA-376 ST_ShapeType 'rightBrace'

RIGHT_BRACKET: Right bracket shape. Corresponds to ECMA-376 ST_ShapeType 'rightBracket'

ROUND_1_RECTANGLE: One round corner rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'round1Rect'

ROUND_2_DIAGONAL_RECTANGLE: Two diagonal round corner rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'round2DiagRect'

ROUND_2_SAME_RECTANGLE: Two same-side round corner rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'round2SameRect'

RIGHT_TRIANGLE: Right triangle shape. Corresponds to ECMA-376 ST_ShapeType 'rtTriangle'

SMILEY_FACE: Smiley face shape. Corresponds to ECMA-376 ST_ShapeType 'smileyFace'

SNIP_1_RECTANGLE: One snip corner rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'snip1Rect'

SNIP_2_DIAGONAL_RECTANGLE: Two diagonal snip corner rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'snip2DiagRect'

SNIP_2_SAME_RECTANGLE: Two same-side snip corner rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'snip2SameRect'

SNIP_ROUND_RECTANGLE: One snip one round corner rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'snipRoundRect'

STAR_10: Ten pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star10'

STAR_12: Twelve pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star12'

STAR_16: Sixteen pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star16'

STAR_24: Twenty four pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star24'

STAR_32: Thirty two pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star32'

STAR_4: Four pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star4'

STAR_5: Five pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star5'

STAR_6: Six pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star6'

STAR_7: Seven pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star7'

STAR_8: Eight pointed star shape. Corresponds to ECMA-376 ST_ShapeType 'star8'

STRIPED_RIGHT_ARROW: Striped right arrow shape. Corresponds to ECMA-376 ST_ShapeType 'stripedRightArrow'

SUN: Sun shape. Corresponds to ECMA-376 ST_ShapeType 'sun'

TRAPEZOID: Trapezoid shape. Corresponds to ECMA-376 ST_ShapeType 'trapezoid'

TRIANGLE: Triangle shape. Corresponds to ECMA-376 ST_ShapeType 'triangle'

UP_ARROW: Up arrow shape. Corresponds to ECMA-376 ST_ShapeType 'upArrow'

UP_ARROW_CALLOUT: Callout up arrow shape. Corresponds to ECMA-376 ST_ShapeType 'upArrowCallout'

UP_DOWN_ARROW: Up down arrow shape. Corresponds to ECMA-376 ST_ShapeType 'upDownArrow'

UTURN_ARROW: U-turn arrow shape. Corresponds to ECMA-376 ST_ShapeType 'uturnArrow'

VERTICAL_SCROLL: Vertical scroll shape. Corresponds to ECMA-376 ST_ShapeType 'verticalScroll'

WAVE: Wave shape. Corresponds to ECMA-376 ST_ShapeType 'wave'

WEDGE_ELLIPSE_CALLOUT: Callout wedge ellipse shape. Corresponds to ECMA-376 ST_ShapeType 'wedgeEllipseCallout'

WEDGE_RECTANGLE_CALLOUT: Callout wedge rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'wedgeRectCallout'

WEDGE_ROUND_RECTANGLE_CALLOUT: Callout wedge round rectangle shape. Corresponds to ECMA-376 ST_ShapeType 'wedgeRoundRectCallout'

FLOW_CHART_ALTERNATE_PROCESS: Alternate process flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartAlternateProcess'

FLOW_CHART_COLLATE: Collate flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartCollate'

FLOW_CHART_CONNECTOR: Connector flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartConnector'

FLOW_CHART_DECISION: Decision flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartDecision'

FLOW_CHART_DELAY: Delay flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartDelay'

FLOW_CHART_DISPLAY: Display flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartDisplay'

FLOW_CHART_DOCUMENT: Document flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartDocument'

FLOW_CHART_EXTRACT: Extract flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartExtract'

FLOW_CHART_INPUT_OUTPUT: Input output flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartInputOutput'

FLOW_CHART_INTERNAL_STORAGE: Internal storage flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartInternalStorage'

FLOW_CHART_MAGNETIC_DISK: Magnetic disk flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartMagneticDisk'

FLOW_CHART_MAGNETIC_DRUM: Magnetic drum flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartMagneticDrum'

FLOW_CHART_MAGNETIC_TAPE: Magnetic tape flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartMagneticTape'

FLOW_CHART_MANUAL_INPUT: Manual input flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartManualInput'

FLOW_CHART_MANUAL_OPERATION: Manual operation flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartManualOperation'

FLOW_CHART_MERGE: Merge flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartMerge'

FLOW_CHART_MULTIDOCUMENT: Multi-document flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartMultidocument'

FLOW_CHART_OFFLINE_STORAGE: Offline storage flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartOfflineStorage'

FLOW_CHART_OFFPAGE_CONNECTOR: Off-page connector flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartOffpageConnector'

FLOW_CHART_ONLINE_STORAGE: Online storage flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartOnlineStorage'

FLOW_CHART_OR: Or flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartOr'

FLOW_CHART_PREDEFINED_PROCESS: Predefined process flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartPredefinedProcess'

FLOW_CHART_PREPARATION: Preparation flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartPreparation'

FLOW_CHART_PROCESS: Process flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartProcess'

FLOW_CHART_PUNCHED_CARD: Punched card flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartPunchedCard'

FLOW_CHART_PUNCHED_TAPE: Punched tape flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartPunchedTape'

FLOW_CHART_SORT: Sort flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartSort'

FLOW_CHART_SUMMING_JUNCTION: Summing junction flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartSummingJunction'

FLOW_CHART_TERMINATOR: Terminator flow shape. Corresponds to ECMA-376 ST_ShapeType 'flowChartTerminator'

ARROW_EAST: East arrow shape.

ARROW_NORTH_EAST: Northeast arrow shape.

ARROW_NORTH: North arrow shape.

SPEECH: Speech shape.

STARBURST: Star burst shape.

TEARDROP: Teardrop shape. Corresponds to ECMA-376 ST_ShapeType 'teardrop'

ELLIPSE_RIBBON: Ellipse ribbon shape. Corresponds to ECMA-376 ST_ShapeType 'ellipseRibbon'

ELLIPSE_RIBBON_2: Ellipse ribbon 2 shape. Corresponds to ECMA-376 ST_ShapeType 'ellipseRibbon2'

CLOUD_CALLOUT: Callout cloud shape. Corresponds to ECMA-376 ST_ShapeType 'cloudCallout'

CUSTOM: Custom shape.

Value

CreateShapeRequest object

Details

Creates a new shape.