LineConnection Object

LineConnection Object

Usage

LineConnection(connectedObjectId = NULL, connectionSiteIndex = NULL)

Arguments

connectedObjectId

(string) The object ID of the connected page element. Some page elements, such as groups, tables, and lines do not have connection sites and therefore cannot be connected to a connector line.

connectionSiteIndex

(integer) The index of the connection site on the connected page element. In most cases, it corresponds to the predefined connection site index from the ECMA-376 standard. More information on those connection sites can be found in the description of the "cnx" attribute in section 20.1.9.9 and Annex H. "Predefined DrawingML Shape and Text Geometries" of "Office Open XML File Formats-Fundamentals and Markup Language Reference", part 1 of ECMA-376 5th edition. The position of each connection site can also be viewed from Slides editor.

Value

LineConnection object

Details

The properties for one end of a Line connection.