Page Object

Page Object

Usage

Page(
  objectId = NULL,
  pageType = NULL,
  pageElements = NULL,
  slideProperties = NULL,
  layoutProperties = NULL,
  notesProperties = NULL,
  masterProperties = NULL,
  revisionId = NULL,
  pageProperties = NULL
)

Arguments

objectId

(string) The object ID for this page. Object IDs used by Page and PageElement share the same namespace.

pageType

(string) The type of the page.

SLIDE: A slide page.

MASTER: A master slide page.

LAYOUT: A layout page.

NOTES: A notes page.

NOTES_MASTER: A notes master page.

pageElements

(list of PageElement objects) The page elements rendered on the page.

slideProperties

(SlideProperties) Slide specific properties. Only set if page_type = SLIDE.

layoutProperties

(LayoutProperties) Layout specific properties. Only set if page_type = LAYOUT.

notesProperties

(NotesProperties) Notes specific properties. Only set if page_type = NOTES.

masterProperties

(MasterProperties) Master specific properties. Only set if page_type = MASTER.

revisionId

(string) Output only. The revision ID of the presentation. Can be used in update requests to assert the presentation revision hasn't changed since the last read operation. Only populated if the user has edit access to the presentation. The revision ID is not a sequential number but an opaque string. The format of the revision ID might change over time. A returned revision ID is only guaranteed to be valid for 24 hours after it has been returned and cannot be shared across users. If the revision ID is unchanged between calls, then the presentation has not changed. Conversely, a changed ID (for the same presentation and user) usually means the presentation has been updated. However, a changed ID can also be due to internal factors such as ID format changes.

pageProperties

(PageProperties) The properties of the page.

Value

Page object

Details

A page in a presentation.