PageElement Object

PageElement Object

Usage

PageElement(
  objectId = NULL,
  size = NULL,
  transform = NULL,
  title = NULL,
  description = NULL,
  elementGroup = NULL,
  shape = NULL,
  image = NULL,
  video = NULL,
  line = NULL,
  table = NULL,
  wordArt = NULL,
  sheetsChart = NULL,
  speakerSpotlight = NULL
)

Arguments

objectId

(string) The object ID for this page element. Object IDs used by google.apps.slides.v1.Page and google.apps.slides.v1.PageElement share the same namespace.

size

(Size) The size of the page element.

transform

(AffineTransform) The transform of the page element. The visual appearance of the page element is determined by its absolute transform. To compute the absolute transform, preconcatenate a page element's transform with the transforms of all of its parent groups. If the page element is not in a group, its absolute transform is the same as the value in this field. The initial transform for the newly created Group is always the identity transform.

title

(string) The title of the page element. Combined with description to display alt text. The field is not supported for Group elements.

description

(string) The description of the page element. Combined with title to display alt text. The field is not supported for Group elements.

elementGroup

(Group) A collection of page elements joined as a single unit.

shape

(Shape) A generic shape.

image

(Image) An image page element.

video

(Video) A video page element.

line

(Line) A line page element.

table

(Table) A table page element.

wordArt

(WordArt) A word art page element.

sheetsChart

(SheetsChart) A linked chart embedded from Google Sheets. Unlinked charts are represented as images.

speakerSpotlight

(SpeakerSpotlight) A Speaker Spotlight.

Value

PageElement object

Details

A visual element rendered on a page.