Recolor Object

Recolor Object

Usage

Recolor(recolorStops = NULL, name = NULL)

Arguments

recolorStops

(list of ColorStop objects) The recolor effect is represented by a gradient, which is a list of color stops. The colors in the gradient will replace the corresponding colors at the same position in the color palette and apply to the image. This property is read-only.

name

(string) The name of the recolor effect. The name is determined from the recolor_stops by matching the gradient against the colors in the page's current color scheme. This property is read-only.

NONE: No recolor effect. The default value.

LIGHT1: A recolor effect that lightens the image using the page's first available color from its color scheme.

LIGHT2: A recolor effect that lightens the image using the page's second available color from its color scheme.

LIGHT3: A recolor effect that lightens the image using the page's third available color from its color scheme.

LIGHT4: A recolor effect that lightens the image using the page's forth available color from its color scheme.

LIGHT5: A recolor effect that lightens the image using the page's fifth available color from its color scheme.

LIGHT6: A recolor effect that lightens the image using the page's sixth available color from its color scheme.

LIGHT7: A recolor effect that lightens the image using the page's seventh available color from its color scheme.

LIGHT8: A recolor effect that lightens the image using the page's eighth available color from its color scheme.

LIGHT9: A recolor effect that lightens the image using the page's ninth available color from its color scheme.

LIGHT10: A recolor effect that lightens the image using the page's tenth available color from its color scheme.

DARK1: A recolor effect that darkens the image using the page's first available color from its color scheme.

DARK2: A recolor effect that darkens the image using the page's second available color from its color scheme.

DARK3: A recolor effect that darkens the image using the page's third available color from its color scheme.

DARK4: A recolor effect that darkens the image using the page's fourth available color from its color scheme.

DARK5: A recolor effect that darkens the image using the page's fifth available color from its color scheme.

DARK6: A recolor effect that darkens the image using the page's sixth available color from its color scheme.

DARK7: A recolor effect that darkens the image using the page's seventh available color from its color scheme.

DARK8: A recolor effect that darkens the image using the page's eighth available color from its color scheme.

DARK9: A recolor effect that darkens the image using the page's ninth available color from its color scheme.

DARK10: A recolor effect that darkens the image using the page's tenth available color from its color scheme.

GRAYSCALE: A recolor effect that recolors the image to grayscale.

NEGATIVE: A recolor effect that recolors the image to negative grayscale.

SEPIA: A recolor effect that recolors the image using the sepia color.

CUSTOM: Custom recolor effect. Refer to recolor_stops for the concrete gradient.

Value

Recolor object

Details

A recolor effect applied on an image.