ReplaceAllShapesWithImageRequest Object

ReplaceAllShapesWithImageRequest Object

Usage

ReplaceAllShapesWithImageRequest(
  containsText = NULL,
  imageUrl = NULL,
  replaceMethod = NULL,
  imageReplaceMethod = NULL,
  pageObjectIds = NULL
)

Arguments

containsText

(SubstringMatchCriteria) If set, this request will replace all of the shapes that contain the given text.

imageUrl

(string) The image URL. The image is fetched once at insertion time and a copy is stored for display inside the presentation. Images must be less than 50MB in size, cannot exceed 25 megapixels, and must be in one of PNG, JPEG, or GIF format. The provided URL can be at most 2 kB in length. The URL itself is saved with the image, and exposed via the Image.source_url field.

replaceMethod

(string) The replace method. Deprecated: use image_replace_method instead. If you specify both a replace_method and an image_replace_method, the image_replace_method takes precedence.

CENTER_INSIDE: Scales and centers the image to fit within the bounds of the original shape and maintains the image's aspect ratio. The rendered size of the image may be smaller than the size of the shape. This is the default method when one is not specified.

CENTER_CROP: Scales and centers the image to fill the bounds of the original shape. The image may be cropped in order to fill the shape. The rendered size of the image will be the same as that of the original shape.

imageReplaceMethod

(string) The image replace method. If you specify both a replace_method and an image_replace_method, the image_replace_method takes precedence. If you do not specify a value for image_replace_method, but specify a value for replace_method, then the specified replace_method value is used. If you do not specify either, then CENTER_INSIDE is used.

IMAGE_REPLACE_METHOD_UNSPECIFIED: Unspecified image replace method. This value must not be used.

CENTER_INSIDE: Scales and centers the image to fit within the bounds of the original shape and maintains the image's aspect ratio. The rendered size of the image may be smaller than the size of the shape. This is the default method when one is not specified.

CENTER_CROP: Scales and centers the image to fill the bounds of the original shape. The image may be cropped in order to fill the shape. The rendered size of the image will be the same as the original shape.

pageObjectIds

(list of string objects) If non-empty, limits the matches to page elements only on the given pages. Returns a 400 bad request error if given the page object ID of a notes page or a notes master, or if a page with that object ID doesn't exist in the presentation.

Value

ReplaceAllShapesWithImageRequest object

Details

Replaces all shapes that match the given criteria with the provided image. The images replacing the shapes are rectangular after being inserted into the presentation and do not take on the forms of the shapes.