Request Object

Request Object

Usage

Request(...)

Arguments

createSlide

(CreateSlideRequest) Creates a new slide.

createShape

(CreateShapeRequest) Creates a new shape.

createTable

(CreateTableRequest) Creates a new table.

insertText

(InsertTextRequest) Inserts text into a shape or table cell.

insertTableRows

(InsertTableRowsRequest) Inserts rows into a table.

insertTableColumns

(InsertTableColumnsRequest) Inserts columns into a table.

deleteTableRow

(DeleteTableRowRequest) Deletes a row from a table.

deleteTableColumn

(DeleteTableColumnRequest) Deletes a column from a table.

replaceAllText

(ReplaceAllTextRequest) Replaces all instances of specified text.

deleteObject

(DeleteObjectRequest) Deletes a page or page element from the presentation.

updatePageElementTransform

(UpdatePageElementTransformRequest) Updates the transform of a page element.

updateSlidesPosition

(UpdateSlidesPositionRequest) Updates the position of a set of slides in the presentation.

deleteText

(DeleteTextRequest) Deletes text from a shape or a table cell.

createImage

(CreateImageRequest) Creates an image.

createVideo

(CreateVideoRequest) Creates a video.

createSheetsChart

(CreateSheetsChartRequest) Creates an embedded Google Sheets chart.

createLine

(CreateLineRequest) Creates a line.

refreshSheetsChart

(RefreshSheetsChartRequest) Refreshes a Google Sheets chart.

updateShapeProperties

(UpdateShapePropertiesRequest) Updates the properties of a Shape.

updateImageProperties

(UpdateImagePropertiesRequest) Updates the properties of an Image.

updateVideoProperties

(UpdateVideoPropertiesRequest) Updates the properties of a Video.

updatePageProperties

(UpdatePagePropertiesRequest) Updates the properties of a Page.

updateTableCellProperties

(UpdateTableCellPropertiesRequest) Updates the properties of a TableCell.

updateLineProperties

(UpdateLinePropertiesRequest) Updates the properties of a Line.

createParagraphBullets

(CreateParagraphBulletsRequest) Creates bullets for paragraphs.

replaceAllShapesWithImage

(ReplaceAllShapesWithImageRequest) Replaces all shapes matching some criteria with an image.

duplicateObject

(DuplicateObjectRequest) Duplicates a slide or page element.

updateTextStyle

(UpdateTextStyleRequest) Updates the styling of text within a Shape or Table.

replaceAllShapesWithSheetsChart

(ReplaceAllShapesWithSheetsChartRequest) Replaces all shapes matching some criteria with a Google Sheets chart.

deleteParagraphBullets

(DeleteParagraphBulletsRequest) Deletes bullets from paragraphs.

updateParagraphStyle

(UpdateParagraphStyleRequest) Updates the styling of paragraphs within a Shape or Table.

updateTableBorderProperties

(UpdateTableBorderPropertiesRequest) Updates the properties of the table borders in a Table.

updateTableColumnProperties

(UpdateTableColumnPropertiesRequest) Updates the properties of a Table column.

updateTableRowProperties

(UpdateTableRowPropertiesRequest) Updates the properties of a Table row.

mergeTableCells

(MergeTableCellsRequest) Merges cells in a Table.

unmergeTableCells

(UnmergeTableCellsRequest) Unmerges cells in a Table.

groupObjects

(GroupObjectsRequest) Groups objects, such as page elements.

ungroupObjects

(UngroupObjectsRequest) Ungroups objects, such as groups.

updatePageElementAltText

(UpdatePageElementAltTextRequest) Updates the alt text title and/or description of a page element.

replaceImage

(ReplaceImageRequest) Replaces an existing image with a new image.

updateSlideProperties

(UpdateSlidePropertiesRequest) Updates the properties of a Slide

updatePageElementsZOrder

(UpdatePageElementsZOrderRequest) Updates the Z-order of page elements.

updateLineCategory

(UpdateLineCategoryRequest) Updates the category of a line.

rerouteLine

(RerouteLineRequest) Reroutes a line such that it's connected at the two closest connection sites on the connected page elements.

Value

Request object

Details

A single kind of update to apply to a presentation.