Response Object

Response Object

Usage

Response(
  createSlide = NULL,
  createShape = NULL,
  createTable = NULL,
  replaceAllText = NULL,
  createImage = NULL,
  createVideo = NULL,
  createSheetsChart = NULL,
  createLine = NULL,
  replaceAllShapesWithImage = NULL,
  duplicateObject = NULL,
  replaceAllShapesWithSheetsChart = NULL,
  groupObjects = NULL
)

Arguments

createSlide

(CreateSlideResponse) The result of creating a slide.

createShape

(CreateShapeResponse) The result of creating a shape.

createTable

(CreateTableResponse) The result of creating a table.

replaceAllText

(ReplaceAllTextResponse) The result of replacing text.

createImage

(CreateImageResponse) The result of creating an image.

createVideo

(CreateVideoResponse) The result of creating a video.

createSheetsChart

(CreateSheetsChartResponse) The result of creating a Google Sheets chart.

createLine

(CreateLineResponse) The result of creating a line.

replaceAllShapesWithImage

(ReplaceAllShapesWithImageResponse) The result of replacing all shapes matching some criteria with an image.

duplicateObject

(DuplicateObjectResponse) The result of duplicating an object.

replaceAllShapesWithSheetsChart

(ReplaceAllShapesWithSheetsChartResponse) The result of replacing all shapes matching some criteria with a Google Sheets chart.

groupObjects

(GroupObjectsResponse) The result of grouping objects.

Value

Response object

Details

A single response from an update.