Shadow Object

Shadow Object

Usage

Shadow(
  type = NULL,
  transform = NULL,
  alignment = NULL,
  blurRadius = NULL,
  color = NULL,
  alpha = NULL,
  rotateWithShape = NULL,
  propertyState = NULL
)

Arguments

type

(string) The type of the shadow. This property is read-only.

transform

(AffineTransform) Transform that encodes the translate, scale, and skew of the shadow, relative to the alignment position.

alignment

(string) The alignment point of the shadow, that sets the origin for translate, scale and skew of the shadow. This property is read-only.

blurRadius

(Dimension) The radius of the shadow blur. The larger the radius, the more diffuse the shadow becomes.

color

(OpaqueColor) The shadow color value.

alpha

(number) The alpha of the shadow's color, from 0.0 to 1.0.

rotateWithShape

(boolean) Whether the shadow should rotate with the shape. This property is read-only.

propertyState

(string) The shadow property state. Updating the shadow on a page element will implicitly update this field to RENDERED, unless another value is specified in the same request. To have no shadow on a page element, set this field to NOT_RENDERED. In this case, any other shadow fields set in the same request will be ignored.

Value

Shadow object

Details

The shadow properties of a page element. If these fields are unset, they may be inherited from a parent placeholder if it exists. If there is no parent, the fields will default to the value used for new page elements created in the Slides editor, which may depend on the page element kind.