SolidFill Object

SolidFill Object

Usage

SolidFill(color = NULL, alpha = NULL)

Arguments

color

(OpaqueColor) The color value of the solid fill.

alpha

(number) The fraction of this color that should be applied to the pixel. That is, the final pixel color is defined by the equation: pixel color = alpha

(color) + (1.0 - alpha)

(background color) This means that a value of 1.0 corresponds to a solid color, whereas a value of 0.0 corresponds to a completely transparent color.

Value

SolidFill object

Details

A solid color fill. The page or page element is filled entirely with the specified color value. If any field is unset, its value may be inherited from a parent placeholder if it exists.