UpdatePageElementsZOrderRequest Object

UpdatePageElementsZOrderRequest Object

Usage

UpdatePageElementsZOrderRequest(pageElementObjectIds = NULL, operation = NULL)

Arguments

pageElementObjectIds

(list of string objects) The object IDs of the page elements to update. All the page elements must be on the same page and must not be grouped.

operation

(string) The Z-order operation to apply on the page elements. When applying the operation on multiple page elements, the relative Z-orders within these page elements before the operation is maintained.

Value

UpdatePageElementsZOrderRequest object

Details

Updates the Z-order of page elements. Z-order is an ordering of the elements on the page from back to front. The page element in the front may cover the elements that are behind it.