Produce configured token

For internal use or for those programming around the Slides API. Returns a token pre-processed with httr::config(). Most users do not need to handle tokens "by hand" or, even if they need some control, slides_auth() is what they need. If there is no current token, slides_auth() is called to either load from cache or initiate OAuth2.0 flow. If auth has been deactivated via slides_deauth(), slides_token() returns NULL.

Usage

ladder_token()

Value

A request object (an S3 class provided by httr).

See also

Examples

if (FALSE) { # \dontrun{
ladder_token()
} # }